Suggested changes to attention implemenation
I have made a few changes, most of which tried to simply the code, but some of which addressed what I think were bugs. Let me know what you think. I was not able to run it so there may well be dimension problems, primarily with things related to decoder_states