Feed the LSTM input with attentions
Feed the LSTM input with attentions.  Below is the important
information (default knowledge) for implement RNN Decoder
```
“feeding the previous step’s attention output as additional input to
the next step of the RNN is somewhat standard. I think that part of the
code came from elsewhere (the TextSum repo) and I wasn’t actually fully
aware of it which is why it wasn’t included in the paper (an oversight
- but that’s why it’s important to release code). I think there’s a
question about it in the Github repo. Or perhaps on the blog post.”
```